refactor(layout): #53 move providers + litellm_config to support/ subpackage (slice 3/3)
Drops src/llmcli/ top-level entry count from 12 → 10 (excluding __pycache__),
giving headroom before the folder-size gate (max 12) fires on the next
routine module add.

git mv preserves history: providers.py → support/providers.py;
litellm_config.py → support/litellm_config.py. New support/__init__.py
holds only a docstring — canonical import paths are
llmcli.support.providers / llmcli.support.litellm_config.

35 import sites updated across 10 files. .importlinter contracts updated
(3 kept, 0 broken). lint-imports passes; ruff clean; pytest 169/169 green.

Closes the 5-finding pre-cascade audit for llmCLI (lyra #1277 framework).
