+++ .importlinter
@@ -6,7 +6,7 @@ include_external_packages = True
 # Layer ordering (top → bottom): higher layers may import lower; lower may NOT import higher.
 #
 #  Layer 3 (high):  cli          — entry point, orchestrates everything
-#  Layer 2 (mid):   daemon, litellm_config — cross-cutting services
+#  Layer 2 (mid):   daemon, support (litellm_config + providers) — cross-cutting services
 #  Layer 1 (low):   engines (impl) — concrete engine backends
 #  Layer 0 (base):  config, engine — pure data/protocol types; no llmcli imports
 #
@@ -17,7 +17,7 @@ name = cli does not import upward (mid layer)
 type = forbidden
 source_modules =
     llmcli.daemon
-    llmcli.litellm_config
+    llmcli.support.litellm_config
 forbidden_modules =
     llmcli.cli
 
@@ -26,7 +26,7 @@ name = mid layer does not import cli
 type = forbidden
 source_modules =
     llmcli.daemon
-    llmcli.litellm_config
+    llmcli.support.litellm_config
     llmcli.engines
 forbidden_modules =
     llmcli.cli
@@ -37,9 +37,9 @@ type = forbidden
 source_modules =
     llmcli.config
     llmcli.engine
-    llmcli.providers
+    llmcli.support.providers
 forbidden_modules =
     llmcli.cli
     llmcli.daemon
-    llmcli.litellm_config
+    llmcli.support.litellm_config
     llmcli.engines
