+++ src/llmcli/support/__init__.py
@@ -1,0 +1,5 @@
+"""Internal-support modules: providers catalog + LiteLLM proxy config.
+
+Public-ish API is re-exported here for convenience but the canonical
+import path is `llmcli.support.providers` / `llmcli.support.litellm_config`.
+"""
